Move param_num to _test_utils
