+++ modelopt/torch/utils/network.py
@@ -26,6 +26,10 @@ from typing import Any, Union
 import torch
 import torch.distributed.fsdp
 import torch.nn as nn
+from torch.nn.modules.batchnorm import _BatchNorm
+from tqdm import tqdm
+
+from .tensor import torch_to
 
 try:
     from torch.distributed.fsdp._state_dict_utils import _convert_to_wrapped_module_name
@@ -37,11 +41,6 @@ except ImportError:
     def _convert_to_wrapped_module_name(name: str) -> str:
         return name
 
-
-from torch.nn.modules.batchnorm import _BatchNorm
-from tqdm import tqdm
-
-from .tensor import torch_to
 
 __all__ = [
     "ModelLike",
@@ -56,7 +55,6 @@ __all__ = [
     "is_parallel",
     "make_divisible",
     "model_to",
-    "param_num",
     "param_num_from_forward",
     "remove_bn",
     "run_forward_loop",
@@ -99,35 +97,6 @@ def get_module_device(module: nn.Module) -> torch.device:
     except StopIteration:
         # For modules without parameters
         return torch.device("cpu")
-
-
-def param_num(network: nn.Module, trainable_only: bool = False, unit=1e6) -> float:
-    """Get the number of parameters of a PyTorch model.
-
-    Args:
-        network: The PyTorch model.
-        trainable_only: Whether to only count trainable parameters. Default is False.
-        unit: The unit to return the number of parameters in. Default is 1e6 (million).
-
-    Returns:
-        The number of parameters in the model in the given unit.
-    """
-    from modelopt.torch.opt.dynamic import DynamicModule
-
-    if isinstance(network, DynamicModule):
-        # NOTE: model.parameters() doesnt consider active_slice so we dont get sorted or trimmed parameters!
-        raise NotImplementedError(
-            "param_num doesn't support DynamicModule. Please use param_num_from_forward instead."
-        )
-    return (
-        sum(
-            p.numel() if not trainable_only or p.requires_grad else 0
-            for mod in network.modules()
-            for p in mod.parameters(recurse=False)
-            if not isinstance(mod, _BatchNorm)
-        )
-        / unit
-    )
 
 
 # TODO: we could also use the same approach as in inference_flops to get the number of params,
